safari-600.7-branch/LayoutTests/fast/dom/NodeList/nodelist-name-getter-properties.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<div id='aWildId'>Wild</div>

<script>
description('Test the JavaScript property descriptor of the name getter.');

var NodeList = document.getElementsByTagName('div');

shouldBe("Object.getOwnPropertyDescriptor(NodeList, 'aWildId').configurable", 'false');
shouldBe("Object.getOwnPropertyDescriptor(NodeList, 'aWildId').enumerable", 'false');
shouldBe("Object.getOwnPropertyDescriptor(NodeList, 'aWildId').writable", 'false');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
